refactor(spdm-controller): extract into dedicated crate
First of several state controllers to be moved out of `crates/api`;
more will follow.

Redfish error conversions to `StateHandlerError` move into
`crates/redfish`.
