+++ crates/api/src/state_controller/external_service_error.rs
@@ -16,39 +16,9 @@
  */
 
 use carbide_dpf::DpfError;
-use carbide_redfish::libredfish::RedfishClientCreationError;
-use libredfish::RedfishError;
 use librms::RackManagerError;
 
 use crate::state_controller::state_handler::{ExternalServiceError, StateHandlerError};
-
-// Keep concrete service client errors at the carbide-api boundary. The
-// state-controller crate deliberately depends only on the opaque
-// ExternalServiceError type.
-
-pub(crate) fn redfish_client_creation_error(
-    error: RedfishClientCreationError,
-) -> StateHandlerError {
-    ExternalServiceError::with_source(
-        "redfish",
-        "create_client",
-        error.to_string(),
-        "redfish_client_creation_error",
-        error,
-    )
-    .into()
-}
-
-pub(crate) fn redfish_error(operation: &'static str, error: RedfishError) -> StateHandlerError {
-    ExternalServiceError::with_source(
-        "redfish",
-        operation,
-        error.to_string(),
-        redfish_operation_metric_label(operation),
-        error,
-    )
-    .into()
-}
 
 pub(crate) fn rack_manager_error(
     operation: &'static str,
@@ -70,12 +40,4 @@ pub(crate) fn dpf_error(error: DpfError) -> StateHandlerError {
 
 pub(crate) fn ufm_error(operation: &'static str, error: eyre::Report) -> StateHandlerError {
     ExternalServiceError::new("ufm", operation, error.to_string(), "ib_fabric_error").into()
-}
-
-fn redfish_operation_metric_label(operation: &'static str) -> &'static str {
-    match operation {
-        "restart" => "redfish_restart_error",
-        "lockdown" => "redfish_lockdown_error",
-        _ => "redfish_other_error",
-    }
 }
