+++ crates/api/src/state_controller/machine/handler/attestation.rs
@@ -30,7 +30,6 @@ use model::machine::{
 use sqlx::PgPool;
 
 use crate::handlers::attestation as attestation_handlers;
-use crate::state_controller::external_service_error::redfish_client_creation_error;
 use crate::state_controller::machine::context::MachineStateHandlerContextObjects;
 use crate::state_controller::state_handler::{
     StateHandlerContext, StateHandlerError, StateHandlerOutcome,
@@ -98,7 +97,7 @@ pub(crate) async fn handle_spdm_trigger_state(
     let redfish_client = redfish_client_pool
         .create_client_from_machine(&mh_snapshot.host_snapshot, db_pool)
         .await
-        .map_err(redfish_client_creation_error)?;
+        .map_err(StateHandlerError::from)?;
 
     let devices_scheduled = attestation_handlers::trigger_attestation(
         db_pool,
