+++ crates/spdm-controller/src/context.rs
@@ -15,13 +15,23 @@
  * limitations under the License.
  */
 
-use crate::state_controller::common_services::CommonStateHandlerServices;
-use crate::state_controller::spdm::metrics::AttestationMetrics;
-use crate::state_controller::state_handler::StateHandlerContextObjects;
+use std::sync::Arc;
+
+use carbide_redfish::libredfish::RedfishClientPool;
+use sqlx::PgPool;
+use state_controller::state_handler::StateHandlerContextObjects;
+
+use crate::metrics::AttestationMetrics;
 
 pub struct SpdmStateHandlerContextObjects {}
 
+#[derive(Clone)]
+pub struct SpdmStateHandlerServices {
+    pub db_pool: PgPool,
+    pub redfish_client_pool: Arc<dyn RedfishClientPool>,
+}
+
 impl StateHandlerContextObjects for SpdmStateHandlerContextObjects {
-    type Services = CommonStateHandlerServices;
+    type Services = SpdmStateHandlerServices;
     type ObjectMetrics = AttestationMetrics;
 }
